:tada: Auto dashboard install using gym cli
