+++ tutorials/rl/001_sandboxes/001_sandboxes.py
@@ -31,8 +31,6 @@ import modal
 import json
 import re
 
-import modal
-
 from modal_training_gym import (
     DeploymentConfig,
     EvalConfig,
