+++ tutorials/tutorial_generator/rl/001_sandboxes.py
@@ -66,8 +66,6 @@ def _imports():
     import json
     import re
 
-    import modal
-
     from modal_training_gym import (
         DeploymentConfig,
         EvalConfig,
